rearrenged some path and screens for UI uniformity, padding issue fixed on pages not in tabs
