+++ src/app/(tabs)/category.tsx
@@ -1,3 +1,4 @@
+import Container from "@/components/UI/Container";
 import { CategoryType } from "@/database/categoriesSchema";
 import useCategories from "@/queries/useCategories";
 import { useRouter } from "expo-router";
@@ -36,28 +37,21 @@ const CategoryList = () => {
   }
 
   return (
-    <View style={styles.container}>
+    <Container>
       {categories && categories.length > 0 ? (
         <FlatList
           data={categories}
           renderItem={renderItem}
           keyExtractor={(item) => item.id.toString()}
-          contentContainerStyle={styles.flatListContent}
         />
       ) : (
         <Text>No categories created yet.</Text>
       )}
-    </View>
+    </Container>
   );
 };
 
 const styles = StyleSheet.create({
-  container: {
-    paddingVertical: 10,
-  },
-  flatListContent: {
-    paddingHorizontal: 16,
-  },
   card: {
     backgroundColor: "#e0e0e0",
     borderRadius: 10,
